server/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pydantic-settings
python-multipart
Pillow
python-dotenv
